[TASK] Remove unused classes and consolidate
