--- Classes/Domain/Model/Dto/MissingIdentifierException.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-
-namespace Pixelant\Interest\Domain\Model\Dto;
-
-/**
- * An exception thrown if an identifier is missing.
- */
-class MissingIdentifierException extends \InvalidArgumentException
-{
-}
