--- Classes/Domain/Model/Dto/Relation.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-
-namespace Pixelant\Interest\Domain\Model\Dto;
-
-/**
- * Representing relations between records.
- */
-class Relation
-{
-
-}
