+++ ext_localconf.php
@@ -2,9 +2,6 @@
 defined('TYPO3_MODE') or die('Access denied.');
 
 (static function () {
-    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['clearCachePostProc'][]
-        = \Pixelant\Interest\Hook\ClearCachePostProc::class . '->clearCachePostProc';
-
     $GLOBALS ['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processCmdmapClass']['interest']
         = \Pixelant\Interest\Hook\ProcessCmdmap::class;
